Created QuickRef; mostly complete; Need polishing
This commit marks the completion of the quick reference documentation for users. Cyber-infrastructure team should make any changes to adapt it to the portal before merging. There may be some sections that need further editing or polishing, but the document contains most essential commands that users need to make and test their plugin. I will continue to work on refining the document in future commits, but for now, this version is a useful resource for our users.